--- library-system/src/main/java/com/maids/librarySystem/LibrarySystemApplication.java
@@ -1,13 +1,0 @@
-package com.maids.librarySystem;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class LibrarySystemApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(LibrarySystemApplication.class, args);
-	}
-
-}
