--- library-system/src/main/java/com/maids/librarySystem/book/models/response/BookResModel.java
@@ -1,19 +1,0 @@
-package com.maids.librarySystem.book.models.response;
-
-import java.io.Serializable;
-
-import lombok.Data;
-
-@Data
-public class BookResModel implements Serializable {
-
-	/**
-	 * 
-	 */
-	private static final long serialVersionUID = -644409175997579663L;
-	private int id;
-	private String author;
-	private String title;
-	private String isbn;
-	private int publicationYear;
-}
