--- library-system/src/main/java/com/maids/librarySystem/book/repositories/BookRepository.java
@@ -1,10 +1,0 @@
-package com.maids.librarySystem.book.repositories;
-
-import org.springframework.data.repository.CrudRepository;
-
-import com.maids.librarySystem.book.entities.Book;
-
-public interface BookRepository extends CrudRepository<Book, Integer>{
-    Book findByIsbn(String isbn);
-
-}
