--- library-system/src/main/java/com/maids/librarySystem/borrowingRecord/models/mapInterface/BorrowingRecordMapper.java
@@ -1,8 +1,0 @@
-package com.maids.librarySystem.borrowingRecord.models.mapInterface;
-
-import org.mapstruct.Mapper;
-
-@Mapper(componentModel = "spring")
-public interface BorrowingRecordMapper {
-
-}
