--- library-system/src/main/java/com/maids/librarySystem/borrowingRecord/services/BorrowingRecordService.java
@@ -1,7 +1,0 @@
-package com.maids.librarySystem.borrowingRecord.services;
-
-public interface BorrowingRecordService {
-	
-	int borrowABook (int bookId,int patronId);
-	int returnABook (int bookId,int patronId);
-}
