--- library-system/src/main/java/com/maids/librarySystem/patron/models/response/PatronResModel.java
@@ -1,18 +1,0 @@
-package com.maids.librarySystem.patron.models.response;
-
-import java.io.Serializable;
-
-import lombok.Data;
-
-@Data
-public class PatronResModel implements Serializable {
-/**
-	 * 
-	 */
-	private static final long serialVersionUID = 7756129405105031822L;
-	
-	private int id;
-	private String name;
-	private String email;
-	private String mobile;
-}
