+++ library-system/src/main/java/com/maids/library_system/book/repositories/BookRepository.java
@@ -1,0 +1,10 @@
+package com.maids.library_system.book.repositories;
+
+import org.springframework.data.repository.CrudRepository;
+
+import com.maids.library_system.book.entities.Book;
+
+public interface BookRepository extends CrudRepository<Book, Integer>{
+    Book findByIsbn(String isbn);
+
+}
