+++ library-system/src/main/resources/application.properties
@@ -1,2 +1,15 @@
+server.port=8088
 spring.application.name=library-system
 server.servlet.context-path=/api
+
+management.endpoints.web.exposure.include=health,info
+
+spring.datasource.url=jdbc:h2:mem:bookdb
+spring.datasource.driverClassName=org.h2.Driver
+spring.datasource.username=sa
+spring.datasource.password=password
+spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
+spring.h2.console.enabled=true
+
+springdoc.swagger-ui.disable-swagger-default-url=true
+springdoc.swagger-ui.path=/doc
